Tabular NN and Cluster problem
Created a tabular neural network, it achieved 100% this is what resulted in the cluster problem

the cluster problem was that deep learning AI was reverse engineering the clustering algorithm and just memorizing that. To fix this, I removed the top 5 variable that enabled the clustering and moved the clustering to after the data split because I learned that due to clustering being an unsupervised method, it can cause a dataleak if before.

This was all done in the test folder and the new csv that were created start with test.